refatoraçao de codigo
