+++ src/app.js
@@ -1,2 +1,2 @@
 require = require('esm')(module /* , options */);
-module.exports = require('./bot.js');
+module.exports = require('./bot/bot.js');
