+++ src/bot/comandos/rolardados/rola_dados.js
