+++ src/bot/const/devmodeserver_roles.js
@@ -1,0 +1,25 @@
+module.exports = {
+  // areas
+  developer: '732607336451014669',
+  datascientist: '732607450204733621',
+  devops: '732607609068060753',
+  infra: '732607758980874290',
+  uxui: '732613392912613378',
+  coach: '732608475049230396',
+  business: '732608526874312754',
+  gestor: '732608722068701235',
+  rh: '732608699218002010',
+  // especialidades
+  fullstack: '732607560221196340',
+  mobile: '732610332950003835',
+  backend: '732607499491868724',
+  frontend: '732607536934682704',
+  // linguagens
+  java: '732607292951887993',
+  javascript: '732607042304344095',
+  python: '732607821333659729',
+  ruby: '732612456014282762',
+  php: '732612770679226521',
+  dotnet: '732608002900754483',
+  sap: '732608366266023957',
+};
