+++ src/bot/schemas/botconfig.json
